set unique galera server-id
Each slave in a mysql replication cluster needs to have a unique
'server-id' set. This is an arbitrary integer, and often people select
1,2,3 etc. For easy deterministic uniqueness without needing to look at
what you've already used, a simple filter has been created to allow for
the hostname to be converted into an integer.

The server ID variable has been added to the my.cnf.j2 template
and the galera_server defaults/main.yml file. The generation of the
server-id variable happens within the `galera-install.yml` play making
it so there is no future impact in role requires.

Change-Id: Id15149e59cb03ab268bb3b7eeaf391334dfc0c01
Fixes-Bug: #1479484